+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+- Add new Testflight testers. [@Rishabh Tayal](https://github.com/RishabhTayal)
 - Fixed app share link. [@Rishabh Tayal](https://github.com/RishabhTayal)
 - Added Edit server url on login. [@Eliottiosdevs](https://github.com/Eliottiosdevs)